move message route action to a message controller
